@@ -1,8 +1,8 @@
-import pLimit = require('p-limit')
+import pLimit from 'p-limit'
 
 export default async (concurrency: number, groups: (() => Promise<void>)[][]) => {
   const limitRun = pLimit(concurrency)
   for (const tasks of groups) {
-    await Promise.all(tasks.map(limitRun))
+    await Promise.all(tasks.map((task) => limitRun(task)))
   }
 }
