@@ -1,18 +1,10 @@
 "use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
 Object.defineProperty(exports, "__esModule", { value: true });
-const pLimit = require("p-limit");
-exports.default = (concurrency, groups) => __awaiter(this, void 0, void 0, function* () {
-    const limitRun = pLimit(concurrency);
+const p_limit_1 = require("p-limit");
+exports.default = async (concurrency, groups) => {
+    const limitRun = p_limit_1.default(concurrency);
     for (const tasks of groups) {
-        yield Promise.all(tasks.map(limitRun));
+        await Promise.all(tasks.map((task) => limitRun(task)));
     }
-});
+};
 //# sourceMappingURL=index.js.map
